+++ src/store/store.ts
@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
 import { authReducer, authApi } from './slices/authSlice';
-import { productReducer } from './slices/productSlice';
+import { productApi, productReducer } from './slices/productSlice';
 import { creditScoreReducer } from './slices/creditScoreSlice';
 import { profileApi } from './slices/profileSlice';
 
@@ -12,8 +12,10 @@ export const store = configureStore({
     creditScore: creditScoreReducer,
     [authApi.reducerPath]: authApi.reducer,
     [profileApi.reducerPath]: profileApi.reducer,
+    [productApi.reducerPath]: productApi.reducer,
   },
-  middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(authApi.middleware, profileApi.middleware),
+  middleware: (getDefaultMiddleware) =>
+    getDefaultMiddleware().concat(authApi.middleware, profileApi.middleware, productApi.middleware),
 });
 
 type RootState = ReturnType<typeof store.getState>;
